+++ kitty/options/parse.py
@@ -11,12 +11,12 @@ from kitty.options.utils import (
     clipboard_control, config_or_absolute_path, copy_on_select, cursor_text_color,
     deprecated_hide_window_decorations_aliases, deprecated_macos_show_window_title_in_menubar_alias,
     deprecated_send_text, disable_ligatures, edge_width, env, font_features, hide_window_decorations,
-    macos_option_as_alt, macos_titlebar_color, optional_edge_width, parse_map, parse_mouse_map,
-    resize_draw_strategy, scrollback_lines, scrollback_pager_history_size, shell_integration,
-    store_multiple, symbol_map, tab_activity_symbol, tab_bar_edge, tab_bar_margin_height,
-    tab_bar_min_tabs, tab_fade, tab_font_style, tab_separator, tab_title_template, titlebar_color,
-    to_cursor_shape, to_font_size, to_layout_names, to_modifiers, url_prefixes, url_style,
-    visual_window_select_characters, window_border_width, window_size
+    macos_option_as_alt, macos_titlebar_color, narrow_symbols, optional_edge_width, parse_map,
+    parse_mouse_map, resize_draw_strategy, scrollback_lines, scrollback_pager_history_size,
+    shell_integration, store_multiple, symbol_map, tab_activity_symbol, tab_bar_edge,
+    tab_bar_margin_height, tab_bar_min_tabs, tab_fade, tab_font_style, tab_separator,
+    tab_title_template, titlebar_color, to_cursor_shape, to_font_size, to_layout_names, to_modifiers,
+    url_prefixes, url_style, visual_window_select_characters, window_border_width, window_size
 )
 
 
@@ -1079,6 +1079,10 @@ class Parser:
     def mouse_hide_wait(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['mouse_hide_wait'] = float(val)
 
+    def narrow_symbols(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
+        for k, v in narrow_symbols(val):
+            ans["narrow_symbols"][k] = v
+
     def open_url_with(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['open_url_with'] = to_cmdline(val)
 
@@ -1331,6 +1335,7 @@ def create_result_dict() -> typing.Dict[str, typing.Any]:
         'exe_search_path': {},
         'font_features': {},
         'kitten_alias': {},
+        'narrow_symbols': {},
         'symbol_map': {},
         'watcher': {},
         'map': [],